node: move modules out into top level
